Add build function functionality
