+++ function_template/function.yml
@@ -1,0 +1,12 @@
+provider:
+  name: faas
+  gateway: http://gateway.openfaas:8080
+
+functions:
+  function_name:
+    lang: dockerfile
+    handler: ./function_name
+    image: registry.docker-registry/function_name
+    environment:
+      fprocess: "python3 supervisor.py"
+      sprocess: "/tmp/user_script.sh"
